refactor: Remove watchconnection-serialization (#304)
* Don't override MessageSerializer

* Don't override MessageSerializer in BatteryStatsSerializer

* Don't override MessageSerializer in BoolSettingSerializer

* Don't override MessageSerializer in IntSettingSerializer

* Make sure message receivers check against path

* Remove MessageSerializer from VersionSerializer

* Remove MessageSerializer from AddedOrUpdatedAppsSerializer

* Remove MessageSerializer from AppIconSerializer

* Remove MessageSerializer from CacheValidationSerializer

* Fix unresolved import

* Update nullability

* Remove MessageSerializer from PackageNameSerializer

* Fix AppCacheUpdateWorker using a MessageHandler

* Remove MessageSerializer from RemovedAppsSerializer

* Remove serialization from appmanager

* Remove serialization from dndsync

* Fix unresolved import

* Update RemovedAppsSerializer nullability

* Remove watchconnection-serialization from version catalog

+++ appmanager/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/appmanager/RemovedAppReceiver.kt
@@ -1,26 +1,28 @@
 package com.boswelja.smartwatchextensions.appmanager
 
 import android.content.Context
+import com.boswelja.watchconnection.common.message.MessageReceiver
 import com.boswelja.watchconnection.common.message.ReceivedMessage
-import com.boswelja.watchconnection.serialization.MessageReceiver
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 
 /**
  * A [MessageReceiver] for receiving [AppList] changes and updating the repository.
  */
-class RemovedAppReceiver : MessageReceiver<RemovedApps>(RemovedAppsSerializer), KoinComponent {
+class RemovedAppReceiver : MessageReceiver(), KoinComponent {
 
     private val repository: WatchAppRepository by inject()
     private val iconRepository: WatchAppIconRepository by inject()
 
-    override suspend fun onMessageReceived(context: Context, message: ReceivedMessage<RemovedApps>) {
-        val watchId = message.sourceUid
-        val packages = message.data.packages
-        repository.delete(watchId, packages)
+    override suspend fun onMessageReceived(context: Context, message: ReceivedMessage<ByteArray?>) {
+        if (message.path == RemovedAppsList) {
+            val packages = message.data?.let { RemovedAppsSerializer.deserialize(it) } ?: return
+            val watchId = message.sourceUid
+            repository.delete(watchId, packages.packages)
 
-        packages.forEach {
-            iconRepository.removeIconFor(watchId, it)
+            packages.packages.forEach {
+                iconRepository.removeIconFor(watchId, it)
+            }
         }
     }
 }
